requirements.txt
python-dotenv
newsapi-python
pandas
sqlalchemy
gspread
google-auth
datetime
Scweet
playwright
feedparser
requests